requirements.txt
aiogram==2.11.2
aiohttp==3.7.3
aioredis==1.3.1
appdirs==1.4.4
async-timeout==3.0.1
asyncpg==0.21.0
attrs==20.3.0
google-trans-new==1.1.9
googletrans==3.0.0
py==1.10.0
pyparsing==2.4.7
python-dotenv==0.15.0
requests==2.25.1
schedule==0.6.0
translate==3.5.0
virtualenv==20.3.1
